Common: split image load/save code into separate cpp per file format.
