+++ Common/CMakeLists.txt
@@ -97,8 +97,10 @@ target_sources(common
     gfx/bitmapdata.cpp
     gfx/bitmapdata.h
     gfx/gfx_def.h
+    gfx/image_bmp.cpp
     gfx/image_file.cpp
     gfx/image_file.h
+    gfx/image_pcx.cpp
     gui/guibutton.cpp
     gui/guibutton.h
     gui/guidefines.h
