+++ Common/gfx/image_file.h
@@ -43,6 +43,15 @@ namespace ImageFile
     // "ext" parameter tells which image format to use.
     inline bool SaveImage(const BitmapData& bmdata, Stream* out, const String& ext)
         { return SaveImage(bmdata, false, nullptr, out, ext); }
+
+    // Load BMP image from the stream
+    PixelBuffer LoadBMP(Stream *in, PixelFormat *src_fmt, RGB *pal);
+    // Load PCX image from the stream
+    PixelBuffer LoadPCX(Stream *in, PixelFormat *src_fmt, RGB *pal);
+    // Save BMP image into the stream
+    bool SaveBMP(const BitmapData &bmp, bool skip_alpha, const RGB *pal, Stream *out);
+    // Save PCX image into the stream
+    bool SavePCX(const BitmapData &bmp, bool skip_alpha, const RGB *pal, Stream *out);
 } // namespace ImageFile
 
 } // namespace Common
