+++ Solutions/Common.Lib/Common.Lib.vcxproj.filters
@@ -527,6 +527,12 @@
     <ClCompile Include="..\..\Common\game\interactions_deprecated.cpp">
       <Filter>Source Files\game</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\Common\gfx\image_bmp.cpp">
+      <Filter>Source Files\gfx</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\Common\gfx\image_pcx.cpp">
+      <Filter>Source Files\gfx</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\Common\ac\audiocliptype.h">
